refactor/AOS: Component Util 패키지 생성
